Split main org file into multiple files
Tangling was really slow when having everything in a single file, thus I
split it up into multiple files. The file structure follows that of the
resulting build artifacts.

Might be worth it structuring it differently to no pollute the classpath
and make files reflect semantics and "way of thinking" rather than the
enforced Java file structure.